Aplikacija/requirements.txt
weaviate-client==4.4b4
